requirements.txt
flask
pytesseract
pillow
requests
beautifulsoup4
openai
